Move shared interface to shared file
